requirements.txt
alembic==1.6.5
blinker==1.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.32
Flask-SQLAlchemy==2.5.1
flask-whooshee==0.8.1
Flask-WTF==0.15.1
greenlet==1.1.1
htmlmin==0.1.12
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsmin==2.2.2
lesscpy==0.15.0
Mako==1.1.4
MarkupSafe==2.0.1
ply==3.11
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
Whoosh==2.7.4
WTForms==2.3.3
xxhash==2.0.2
